Big Changes
Added a Dialogue Manager, with Inky as a basis, some more assets into the scene,  an event that prompts a scene change, a better camera (it rotates with the player now), and better movement (rb.moveposition is world based, so i changed it to transform.position. I was worried at first that the collisions wouldn't work but its fine). Next I'll fix up the UI/create dialogue, and work on the next scene