Updated Fixes
Okay! Added Raycasting for NPCs (things you can interact with), a GameManager (for saving the game each time a new scene is loaded, and for quitting the game (in the main menu and last scene)), and fixed the Characters script by adding a generic movement method, as well as removing Bruce's (he has his own movement method in his script).

Also fixed an issue with the AudioManager, where on scene change it wouldn't play audio because the game object was destroyed. Now it works when playing from the Main Menu all the way through the last scene. It won't work on individual scenes (as in starting from, say, scene 2) but I don't think it's an issue.